Rag/vision_rag/requirements.txt
cohere==5.16.1
openai==1.96.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Pillow==11.3.0
numpy==2.3.1
requests==2.32.4
langchain==0.3.26
langchain-openai==0.3.28
langchain-experimental==0.3.4
langchain-core==0.3.68
pymupdf==1.26.3
pdf2image==1.17.0
poppler-utils==0.1.0
httpx==0.28.1
pydantic==2.11.7
google-genai
pgvector==0.4.1
